Backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
django==5.2.5
django-cors-headers==4.7.0
django-stubs==5.2.2
django-stubs-ext==5.2.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
djangorestframework-stubs==3.16.2
dotenv==0.9.9
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mypy==1.17.1
mypy-extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
pip==24.0
platformdirs==4.3.8
pyjwt==2.10.1
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
ruff==0.12.9
setuptools==65.5.1
sqlparse==0.5.3
types-mysqlclient==2.2.0.20250516
types-pyyaml==6.0.12.20250809
types-requests==2.32.4.20250809
typing-extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uv==0.8.11
wheel==0.45.1
